Moved api.js to the new api directory.
